requirements.txt
pymysql
python-dotenv
customtkinter
Flask
flask-cors
Pillow
